requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
pygame-ce==2.5.1
PyTMX==3.32
requests==2.32.3
urllib3==2.2.3